+++ R/utils.R
@@ -18,7 +18,7 @@ is.uy4326 <- function(x){
   y$geometry <- NULL
   x1 <- sf::st_join(x, y, left = F)
   x2 <- (nrow(x1) / nrow(x)) * 100
-  return(ifelse(x2 == 0, "Your object have not any geometry in Uruguay... Try 32721 or 5381",
+  return(ifelse(x2 == 0, "Your object have not any geometry in Uruguay... Try 5381, 5382 or 32721",
                 glue::glue("Your object have {x2}% of matches with Ururguay CRS 4326")))
 }
 
@@ -42,7 +42,7 @@ is.uy32721 <- function(x){
   y$geometry <- NULL
   x1 <- sf::st_join(x, y, left = F)
   x2 <- (nrow(x1) / nrow(x)) * 100
-  return(ifelse(x2 == 0, "Your object have not any geometry in Uruguay... Try 4326 or 5381",
+  return(ifelse(x2 == 0, "Your object have not any geometry in Uruguay... Try 4326, 5381 or 5382",
                 glue::glue("Your object have {x2}% of matches with Ururguay CRS 32721")))
 }
 
@@ -66,6 +66,30 @@ is.uy5381 <- function(x){
   y$geometry <- NULL
   x1 <- sf::st_join(x, y, left = F)
   x2 <- (nrow(x1) / nrow(x)) * 100
-  return(ifelse(x2 == 0, "Your object have not any geometry in Uruguay... Try 4326 or 32721",
+  return(ifelse(x2 == 0, "Your object have not any geometry in Uruguay... Try 4326, 5382 or 32721",
                 glue::glue("Your object have {x2}% of matches with Ururguay CRS 5381")))
+}
+
+#' This function test if an 'sf' object match with Uruguay at crs = 5382.
+#' @param x An 'sf' object with the same crs as the homonym parameter
+#' @keywords CRS sf Uruguay
+#' @return logical value based in crs parameter of the sf object
+#' @importFrom sf st_join st_crs
+#' @importFrom glue glue
+#' @export
+#' @examples
+#'\donttest{
+#' is.uy5382(load_geouy("Peajes"))
+#'}
+
+is.uy5382 <- function(x){
+  md <- geouy::metadata 
+  if (!is(x, "sf")) stop("The object you want to process is not class sf")
+  y <- geouy::load_geouy("Uruguay", crs = 5382)
+  if (st_crs(x) != st_crs(y)) stop("The object you want to process is not at CRS 5382")
+  y$geometry <- NULL
+  x1 <- sf::st_join(x, y, left = F)
+  x2 <- (nrow(x1) / nrow(x)) * 100
+  return(ifelse(x2 == 0, "Your object have not any geometry in Uruguay... Try 4326, 5381 or 32721",
+                glue::glue("Your object have {x2}% of matches with Ururguay CRS 5382")))
 }