+++ NEWS.md
@@ -1,14 +1,15 @@
 # log history of geouy package development
 
 -------------------------------------------------------
-## geouy v0.1.8 (2020-03-04)
+## geouy v0.1.8 (2020-03-17)
 
 * Add to 'tiles_ide_uy()' function options to download .jpg in addition to .tif
 * Add more testthat for 'tiles_ide_uy()' function
 * Add zip format to 'load_goeuy()'
 * Add MVOTMA datasets to metadata, i.e.: Ambientes acuaticos, Areas protegidas, Batimetria, Secciones catastrales, Padrones rurales y urbanos, Secciones policiales, Playas and Cuencas hidrograficas in its 5 nivels.
 * Add to tiles_ide_uy the posibility for Montevideo tiles with urban = TRUE give orthophotos with 10cm per pixel 
-* Solved bug with raster::mosaic for combination of multiple tiles to cover a bbox.  
+* Solved bug with raster::mosaic for combination of multiple tiles to cover a bbox.
+* Add a complementary function for 5382 crs evaluation
 
 ## geouy v0.1.7 (2020-02-26)